Implement repocache-adm command line tool
This command line tool is intended for managing a gbp repository cache.
It will contain functionality for e.g. checking, cleaning up and
repairing the repository cache.

The tool is supposed to be runnable on a live system.

This initial implementation only contains a rudimentary 'stat' command
for printing repocache status/statistics.

Change-Id: I8607fa2c60241e02ef7d8c04c8223f6136ac0943
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>